pyproject.toml
[tool.poetry]
name = "drffuncserialize"
version = "0.1.0"
description = ""
authors = ["Grégoire Deveaux <ggdx@laposte.net>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.14.0"
Django = "^4.1.1"
djangorestframework-camel-case = "^1.3.0"


[tool.poetry.group.dev.dependencies]
pytest-django = "^4.5.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "drffuncserialize.settings"
